Add trollleap target
In order to allow a giant troll face to leap across the screen, this
commit refactors the logic in the project to allow sharing the command
stuff. It allows a new class to conform to the Commandable protocol
which lets it do a lot of the same stuff. It uses defaults from the new
class to decide which images to leap across the screen.